+++ e2e/fixtures/components/is-string/is-string.js
@@ -1,0 +1,14 @@
+/** @flow */
+
+/**
+ * detemines whether `str` is a string.
+ * @name isString
+ * @param {*} val
+ * @returns {boolean}
+ * @example
+ *  isString(3) // => false
+ *  isString('') // => true
+ */
+export default function isString(val: string): boolean {
+  return typeof val === 'string';
+}